use annotations overlay in WorldImpl + implement @ExtensionPriority
With this commit, we also no longer create an `AnnotationSet` if it
isn't needed on certain places. The annotation overlay knows
if annotations are overlaid for given declaration, and if they
aren't, some queries for annotations can be answered without
creating the `AnnotationSet`. This could be done more thoroughly;
for now, it's just about figuring out if an annotation is present,
which the `WorldImpl` does very often.