use the actual CDI Build Compatible Extension proposed API
Given that the Build Compatible Extension API present in CDI
is in the `jakarta` namespace, while ArC is still on `javax`,
we transform the `jakarta` API into `javax` for now. That's
relatively straightforward, so it's done in a sync script.
This sync script is also added in this commit, and will help
with staying up to date with changes in the CDI API.

Clearly this can't be used anywhere, but for development purpose,
it's the easiest way forward. The entire `cdi-lite-ext-api` module
should be removed in the future anyway.